Corrected small mistake with cnt_m in various codes
cnt_m is used to count the modulators and should be incremented
for each channel in mod_Ch. It was previously updated ONLY if the
channel was a modulator which was also not a receiver.
Although we want to consider modulators which are not receivers,
we should count also those which are receivers in order to move
ahead in mod_Ch